feat: Added RGB to LUV conversion, beginning of density function approximation using kernels and data directory
Added a simple conversion of RGB images to CIELUV images, because euclid distance is a natural metric in LUV.
Added a density function approximation using a gaussian kernel. For now the computation time of using the kernel is too big, so changes must be made.
Created data directory and added two images to it, one 512x512 and one 256x256. Not sure if the image size is making a problem with the kernel or if it is something else.