feat: Implemented first iteration of the clustering function, as well as plotted out the persistance diagram
Implemented the full clustering function, which returns the disjoint-set data structure and the births and deaths of all function values.
Used the gudhi library to plot out the persistance diagram.
Changed the UnionFind class to fit the newly implemented function.
Changed the "small" 16x16 image to have some noise. It is currently used for testing as any larger images brick the algorithm. The function should be changed in the future to work with a smaller subset of elements to save on computation time.