feat: Added centroid and cluster calculations
Added centroid calculations by grabbing the exact number of clusters found by the algorithm from the UnionFind data structure.
After that, for each point of the image, we decide to which cluster it belongs.
We also decide a treshold of elements belonging to a cluster needed for that cluster to actually take a color, having the clusters with less than desirable amount of elements stay colorless.
Currently there is an issue concerning returning previously removed unmarked points and deciding their cluster belongings. Needs to be fixed ASAP.